add: mga 2 mainnet deploymnet
