--- deploy/07_giveaway/18_deploy_multi_giveaway_1.ts
@@ -1,20 +1,0 @@
-import {DeployFunction} from 'hardhat-deploy/types';
-import {HardhatRuntimeEnvironment} from 'hardhat/types';
-
-const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const {deployments, getNamedAccounts} = hre;
-  const {deploy} = deployments;
-  const {deployer, multiGiveawayAdmin} = await getNamedAccounts();
-
-  await deploy('Multi_Giveaway_1', {
-    contract: 'MultiGiveaway',
-    from: deployer,
-    log: true,
-    skipIfAlreadyDeployed: true,
-    args: [multiGiveawayAdmin, '0x0000000000000000000000000000000000000000'], // admin, trustedForwarder
-  });
-};
-export default func;
-func.tags = ['Multi_Giveaway_1', 'Multi_Giveaway_1_deploy'];
-func.dependencies = [];
-func.skip = async () => true;
